Update Maps - most up to date.py
Fixed bug, loading files could sometime re-set the game regardless. Also the player no longer has to manually close the tkinter window after saving a file. However, they still need to do this for 1 of the windows due to the not quite removed label system. Option to load a game from a previous save has also been added to the in game (ie. not start) menu.