Update Map.py
on line 299 i made a function called spell_image(). This function is called when an attack gif is needed, but as this bit of code is still in testing and this is the only gif i have it is used for melee attacks as well. latter i will change it's name to the class of spells that it represents.

on line 311 i made a function called attackgif. This function was made for testing purposes. I'll will finish it when the shop code and items are added and a have the gifs for the items and spells.

on line 327 i created an array called spell. This was purely for testing and will be removed latter.

on line 328 i created a variable called weapons. again this was just for testing and will be removed when the shop code is added.

on line 361 i called the attackgif function.