Update Shop (final).py
New features:
Players can now only by items that there class should be able to purchase. System only shows items which players can purchase and checks if they try to buy one not shown.
All of \shop, \help and the error message if you type in something not recognised are sent to the pygame window instead of using print statements.