Syntax
New features:
Added healing square (40,40)
Bug fixes:
Fixed (some of?) the syntax errors which happened because people edited this in the browser. Open it and run it in Python - we all make mistakes but there is no excuse for not checking.

PS: Adam, not sure what happened to your boss push, but I have added it in with this push as well.